Docker Container Setup
The flask app is now containerized in Docker and we are able transfer it easily along with all of its dependencies. nginx and uwsgi are now being used as the web and app servers. We now don't have to worry about downloading every single dependency on everyone's development machines. The only thing everyone needs to have installed is Docker. The command to test the application is now "docker-compose up."